+++ machine/uart16750.h
@@ -5,7 +5,7 @@
 
 #include <stdint.h>
 
-extern volatile uint8_t* uart16750;
+extern volatile uint32_t* uart16750;
 
 void uart16750_putchar(uint8_t ch);
 int uart16750_getchar();
